+++ src/charts/CategoricalBarChart/index.tsx
@@ -11,7 +11,7 @@ import {
 import { getInstanceByDom, init, use } from "echarts/core";
 import { CanvasRenderer } from "echarts/renderers";
 import { useRef, useEffect } from "react";
-import type { UIDDetails } from "../../sample-data/interfaces";
+import type { UIDDetails } from "../../utils";
 import type { BarSeriesOption } from "echarts/charts";
 import type {
   TitleComponentOption,
